feature: combine get by id in one getById(Integer id) and make HistoryManager historyManager as field of class InMemoryTaskManager
