+++ src/main/java/HistoryManager.java
@@ -1,6 +1,8 @@
+import java.util.List;
+
 public interface HistoryManager {
     //Помечать задачи, как просмотренные
     void addTask(Task task);
     //Возвращать список просмотренных задач
-    void getHistory();
+    List<Task> getHistory();
 }
